Add files via upload
The codebase for the Customized Face Generator is a powerful tool for enabling users to create personalized and expressive faces, with applications spanning virtual reality, gaming, entertainment, personalization, and beyond. Its technical prowess and user-centric features make it a valuable asset in the realm of computer vision and image generation.